Everything, except parser

+++ pr1/inc/color.hpp
@@ -1,14 +1,13 @@
 #pragma once
-#include "types.hpp"
 
 #pragma pack(push, 1)
 class Color {
   public:
 	Color();
 	Color(const Color &);
-	Color(byte r, byte g, byte b);
+	Color(float r, float g, float b);
 	~Color() = default;
 
-	byte r, g, b;
+	float r, g, b;
 };
 #pragma pack(pop)
